+++ lineLineIntersec.c
@@ -39,7 +39,7 @@
 void	lineLineIntersec(point_t*, point_t*, point_t*, point_t*, uint32_t*, point_t*);
 
 void	lineLineIntersec(point_t* p1, point_t* p2, point_t* q1, point_t* q2, uint32_t* i, point_t* isect){
-	double		d;
+	float		d;
 
 	*i = 0;
 	d = (p1->r - p2->r) * (q1->z - q2->z) - (p1->z - p2->z) * (q1->r - q2->r);
